pyproject.toml
[tool.poetry]
name = "dunderbot"
version = "0.0.1"
description = "The dunderbot"
authors = ["perjonasam"]

[tool.poetry.dependencies]
python = "^3.7"
pandas = "^0"
jupyter = "^1.0.0"
ccxt = "^1.30.85"
stable-baselines = "^2.10.0"
tensorflow = "^1"
pyyaml = "^5.3.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
